lnst.Devices.RemoteDevice: support remote attribute setting
We need to introduce an attribute that tells us whether the RemoteDevice
object has been initialized. After that, only pass through to the rpc call
those attributes that already exist on the wrapped device. If this turns out
to be too big limitation, it can be lifted in the future.

Signed-off-by: Jiri Benc <jbenc@redhat.com>